requirements.txt
ldap3<=2.0
PyYAML
